requirements.txt
fastapi
uvicorn[standard]
google-generativeai>=0.9.0
